do a class implementation in the CPP file
